--- flutter_01.log
@@ -1,71 +1,0 @@
-Flutter crash report.
-Please report a bug at https://github.com/flutter/flutter/issues.
-
-## command
-
-flutter --no-color run --machine --track-widget-creation --device-id=chrome --start-paused --dart-define=flutter.inspector.structuredErrors=true lib\homeaddress.dart
-
-## exception
-
-StateError: Bad state: No context with the running Dart application.
-
-```
-#0      RemoteDebuggerExecutionContext.id (package:dwds/src/debugging/execution_context.dart:55:7)
-<asynchronous suspension>
-#1      AppInspector._getExtensionRpcs (package:dwds/src/debugging/inspector.dart:538:20)
-<asynchronous suspension>
-#2      AppInspector._initialize (package:dwds/src/debugging/inspector.dart:97:34)
-<asynchronous suspension>
-#3      AppInspector.initialize (package:dwds/src/debugging/inspector.dart:160:5)
-<asynchronous suspension>
-#4      ChromeProxyService.createIsolate (package:dwds/src/services/chrome_proxy_service.dart:217:18)
-<asynchronous suspension>
-```
-
-## flutter doctor
-
-```
-[✓] Flutter (Channel stable, 2.8.1, on Microsoft Windows [Version 10.0.19044.1526], locale en-US)
-    • Flutter version 2.8.1 at C:\Users\ryank\OneDrive\Documents\flutter
-    • Upstream repository https://github.com/flutter/flutter.git
-    • Framework revision 77d935af4d (3 months ago), 2021-12-16 08:37:33 -0800
-    • Engine revision 890a5fca2e
-    • Dart version 2.15.1
-
-[✓] Android toolchain - develop for Android devices (Android SDK version 33.0.0-rc1)
-    • Android SDK at C:/Users/ryank/AppData/Local/Android/Sdk
-    • Platform android-31, build-tools 33.0.0-rc1
-    • ANDROID_HOME = C:/Users/ryank/AppData/Local/Android/Sdk
-    • Java binary at: C:\Program Files\Android\Android Studio\jre\bin\java
-    • Java version OpenJDK Runtime Environment (build 11.0.10+0-b96-7249189)
-    • All Android licenses accepted.
-
-[✓] Chrome - develop for the web
-    • Chrome at C:\Program Files (x86)\Google\Chrome\Application\chrome.exe
-
-[✓] Android Studio (version 2020.3)
-    • Android Studio at C:\Program Files\Android\Android Studio
-    • Flutter plugin can be installed from:
-      🔨 https://plugins.jetbrains.com/plugin/9212-flutter
-    • Dart plugin can be installed from:
-      🔨 https://plugins.jetbrains.com/plugin/6351-dart
-    • Java version OpenJDK Runtime Environment (build 11.0.10+0-b96-7249189)
-
-[✓] IntelliJ IDEA Community Edition (version 2021.2)
-    • IntelliJ at C:\Program Files\JetBrains\IntelliJ IDEA Community Edition 2021.1.3
-    • Flutter plugin can be installed from:
-      🔨 https://plugins.jetbrains.com/plugin/9212-flutter
-    • Dart plugin can be installed from:
-      🔨 https://plugins.jetbrains.com/plugin/6351-dart
-
-[✓] VS Code (version 1.63.2)
-    • VS Code at C:\Users\ryank\AppData\Local\Programs\Microsoft VS Code
-    • Flutter extension can be installed from:
-      🔨 https://marketplace.visualstudio.com/items?itemName=Dart-Code.flutter
-
-[✓] Connected device (2 available)
-    • Chrome (web) • chrome • web-javascript • Google Chrome 99.0.4844.51
-    • Edge (web)   • edge   • web-javascript • Microsoft Edge 98.0.1108.50
-
-• No issues found!
-```
