+++ lib/router.dart
@@ -1,4 +1,5 @@
-import 'package:aura/main_tab_bar.dart';
+import 'package:aura/view/tabs/main_tab_bar.dart';
+import 'package:aura/view/setting_screen.dart';
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 
@@ -6,17 +7,18 @@ final router = GoRouter(
     initialLocation: "/tabs/map",
     urlPathStrategy: UrlPathStrategy.path,
     routes: <GoRoute>[
-  GoRoute(
-      path: "/tabs/:tabName",
-      builder: (BuildContext context, GoRouterState state) =>
-          MainTabBar(key: state.pageKey, currentTab: state.params['tabName']!)),
-  // GoRoute(
-  //     path: "/settings",
-  //     builder: (BuildContext context, GoRouterState state) => const SettingsScreen()),
-  // GoRoute(
-  //     path: "/notifications",
-  //     builder: (BuildContext context, GoRouterState state) => const NotificationsScreen()),
-]
+      GoRoute(
+          path: "/tabs/:tabName",
+          builder: (BuildContext context, GoRouterState state) => MainTabBar(
+              key: state.pageKey, currentTab: state.params['tabName']!)),
+      GoRoute(
+          path: "/settings",
+          builder: (BuildContext context, GoRouterState state) =>
+              SettingsScreen()),
+      // GoRoute(
+      //     path: "/notifications",
+      //     builder: (BuildContext context, GoRouterState state) => const NotificationsScreen()),
+    ]
     // errorPageBuilder: (context, state) => MaterialPage<void>(
     //   key: state.pageKey,
     //   child: ErrorPage(state.error),
