+++ lib/view/tabs/main_tab_bar.dart
@@ -1,13 +1,11 @@
 import 'package:flutter/material.dart';
-import 'package:aura/view/community/community_tab.dart';
-import 'package:aura/view/map/map_tab.dart';
-import 'package:aura/view/news/news_tab.dart';
+import 'package:aura/view/tabs/community/community_tab.dart';
+import 'package:aura/view/tabs/map/map_tab.dart';
+import 'package:aura/view/tabs/news/news_tab.dart';
 import 'package:go_router/go_router.dart';
 /* Navigation Tab Bar */
 
-const tabs = {
-  0: "community", 1: "map", 2: "news"
-};
+const tabs = {0: "community", 1: "map", 2: "news"};
 
 class MainTabBar extends StatefulWidget {
   final String currentTab;
@@ -19,7 +17,6 @@ class MainTabBar extends StatefulWidget {
 }
 
 class _MainTabBarState extends State<MainTabBar> {
-
   @override
   Widget build(BuildContext context) {
     int selectedIndex = tabs.entries
@@ -27,11 +24,25 @@ class _MainTabBarState extends State<MainTabBar> {
         .key;
 
     void changeTab(String newTabName) {
-      print(newTabName);
       GoRouter.of(context).go("/tabs/$newTabName");
     }
 
     return Scaffold(
+        appBar: AppBar(
+          title: const Text('Aura'),
+          backgroundColor: Colors.blue,
+          actions: <Widget>[
+            IconButton(
+              icon: const Icon(
+                Icons.settings,
+                color: Colors.black54,
+              ),
+              onPressed: () {
+                GoRouter.of(context).push("/settings");
+              },
+            )
+          ],
+        ),
         body: IndexedStack(
           index: selectedIndex,
           children: const <Widget>[
